requirements.txt
anthropic>=0.96.0
agno>=2.5.16
sqlalchemy>=2.0.0
pandas>=2.0.0
scipy>=1.13.0
streamlit>=1.40.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tenacity>=9.0.0
loguru>=0.7.0
apscheduler>=3.10.0
requests>=2.31.0
pyyaml>=6.0.0
